move xlib/main.c video fxn to xlib/video.c
